+++ src/main/java/top/theillusivec4/curios/server/command/CuriosCommand.java
@@ -23,10 +23,15 @@ import com.mojang.brigadier.Command;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import java.util.Map;
+import java.util.Set;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.EntityArgument;
+import net.minecraft.commands.arguments.item.ItemArgument;
+import net.minecraft.commands.arguments.item.ItemInput;
+import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.ItemStack;
@@ -35,6 +40,7 @@ import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.type.inventory.ICurioStacksHandler;
 import top.theillusivec4.curios.common.network.NetworkHandler;
 import top.theillusivec4.curios.common.network.server.sync.SPacketSyncCurios;
+import top.theillusivec4.curios.common.slottype.SlotTypeManager;
 
 public class CuriosCommand {
 
@@ -42,6 +48,35 @@ public class CuriosCommand {
 
     LiteralArgumentBuilder<CommandSourceStack> curiosCommand = Commands.literal("curios")
         .requires(player -> player.hasPermission(2));
+
+    curiosCommand.then(Commands.literal("list").executes(context -> {
+      Map<String, Set<String>> map = SlotTypeManager.getIdsToMods();
+
+      for (Map.Entry<String, Set<String>> entry : map.entrySet()) {
+        context.getSource().sendSuccess(
+            new TextComponent(entry.getKey() + " - " + String.join(", ", entry.getValue())), false);
+      }
+      return Command.SINGLE_SUCCESS;
+    }));
+
+    curiosCommand.then(Commands.literal("replace").then(
+        Commands.argument("slot", CurioArgumentType.slot()).then(
+            Commands.argument("index", IntegerArgumentType.integer()).then(
+                Commands.argument("player", EntityArgument.player()).then(
+                    Commands.literal("with").then(
+                        Commands.argument("item", ItemArgument.item()).executes(
+                            context -> replaceItemForPlayer(context.getSource(),
+                                EntityArgument.getPlayer(context, "player"),
+                                CurioArgumentType.getSlot(context, "slot"),
+                                IntegerArgumentType.getInteger(context, "index"),
+                                ItemArgument.getItem(context, "item"))).then(
+                            Commands.argument("count", IntegerArgumentType.integer()).executes(
+                                context -> replaceItemForPlayer(context.getSource(),
+                                    EntityArgument.getPlayer(context, "player"),
+                                    CurioArgumentType.getSlot(context, "slot"),
+                                    IntegerArgumentType.getInteger(context, "index"),
+                                    ItemArgument.getItem(context, "item"),
+                                    IntegerArgumentType.getInteger(context, "count"))))))))));
 
     curiosCommand.then(Commands.literal("set").then(
         Commands.argument("slot", CurioArgumentType.slot()).then(
@@ -94,6 +129,23 @@ public class CuriosCommand {
                 EntityArgument.getPlayer(context, "player")))));
 
     dispatcher.register(curiosCommand);
+  }
+
+  private static int replaceItemForPlayer(CommandSourceStack source, ServerPlayer player,
+                                          String slot, int index, ItemInput item)
+      throws CommandSyntaxException {
+    return replaceItemForPlayer(source, player, slot, index, item, 1);
+  }
+
+  private static int replaceItemForPlayer(CommandSourceStack source, ServerPlayer player,
+                                          String slot, int index, ItemInput item, int count)
+      throws CommandSyntaxException {
+    ItemStack stack = item.createItemStack(count, false);
+    CuriosApi.getCuriosHelper().setEquippedCurio(player, slot, index, stack);
+    source.sendSuccess(
+        new TranslatableComponent("commands.curios.replace.success", slot, player.getDisplayName(),
+            stack.getDisplayName()), true);
+    return Command.SINGLE_SUCCESS;
   }
 
   private static int setSlotsForPlayer(CommandSourceStack source, ServerPlayer playerMP,